data_pipeline/requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
pandas
numpy
scikit-learn
imbalanced-learn
xgboost
torch
transformers
sentencepiece
shap
jupyter
python-dotenv
datasets
accelerate
flask
flask-cors